Adapt document model to work with BinaryCrossEntropy loss

+++ src/datasets.py
@@ -103,7 +103,7 @@ class BinarySpoilerDataset(torch.utils.data.Dataset):
     def __getitem__(self, index):
         return (
             self.saved_data[str(index)],
-            torch.tensor(self.labels[index], dtype=torch.long)
+            torch.tensor(self.labels[index], dtype=torch.float)
         )
 
     def to_feature(self, text) -> TvTropesFeature:
@@ -167,7 +167,7 @@ class TokenSpoilerDataset(BinarySpoilerDataset):
         return TokenFeature(
             token_ids=token_ids,
             sentence_ids=sentence_ids,
-            labels=torch.tensor(spoiler_bools),
+            labels=torch.tensor(spoiler_bools, dtype=torch.float),
         )
 
 
@@ -194,7 +194,7 @@ class PaddedBatch:
             self.labels = torch.stack(transposed[1], 0)
         else:
             size = max(len(el) for el in transposed[1])
-            self.labels = torch.zeros(len(transposed[1]), size, dtype=torch.long)
+            self.labels = torch.zeros(len(transposed[1]), size, dtype=torch.float)
             for i, el in enumerate(transposed[1]):
                 self.labels[i][:len(el)] = el
